% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings_assign_types.R
\name{assign_palm_building_type}
\alias{assign_palm_building_type}
\title{Assign PALM building types using WSF Evolution data}
\usage{
assign_palm_building_type(buildings, wsf)
}
\arguments{
\item{buildings}{An \code{sf} object with building geometries.}

\item{wsf}{A \code{terra::SpatRaster} with WSF Evolution data.}
}
\value{
The input \code{sf} object with added columns
\code{year_max} and \code{palm_type}.
}
\description{
Assigns PALM building type classes based on ALKIS \code{function}
codes and maximum WSF Evolution construction year values
extracted per building footprint.
}
\examples{
\dontrun{
bld <- process_building_vectors(buildings, aoi)$buildings
wsf <- download_wsf_data(aoi)

bld <- assign_palm_building_type(bld, wsf)
}

}
