% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_buildings_to_palm.R
\name{classify_buildings_to_palm}
\alias{classify_buildings_to_palm}
\title{Assign PALM building types using WSF Evolution data}
\usage{
classify_buildings_to_palm(buildings, wsf)
}
\arguments{
\item{wsf}{A \code{terra::SpatRaster} with WSF Evolution data.}

\item{data}{An \code{sf} object with building geometries.}
}
\value{
The input \code{sf} object with added columns
\code{year_max} and \code{palm_type}.
}
\description{
Assigns PALM building type classes based on ALKIS \code{function}
codes and maximum WSF Evolution construction year values
extracted per building footprint.
}
\examples{
\dontrun{
bld <- process_building_vectors(buildings, aoi)$buildings
wsf <- download_wsf_data(aoi)

bld <- assign_palm_building_type(bld, wsf)
}

}
