% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_buildings.R
\name{.add_sequential_id}
\alias{.add_sequential_id}
\title{Add sequential IDs to building features}
\usage{
.add_sequential_id(x, id_col = "ID")
}
\description{
Add sequential IDs to building features
}
\keyword{internal}
