% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_csd_config.R
\name{.auto_discover_input_files}
\alias{.auto_discover_input_files}
\title{Auto-discover PALM input files based on naming patterns}
\usage{
.auto_discover_input_files(input_root_path, files)
}
\description{
Auto-discover PALM input files based on naming patterns
}
\keyword{internal}
