% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_csd_config.R
\name{.build_csd_list}
\alias{.build_csd_list}
\title{Build structured CSD configuration list}
\usage{
.build_csd_list(args, files)
}
\description{
Build structured CSD configuration list
}
\keyword{internal}
