% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_buildings.R
\name{.classify_palm_type}
\alias{.classify_palm_type}
\title{Classify buildings into PALM types}
\usage{
.classify_palm_type(function_code, year_max)
}
\description{
Assigns PALM building types (1-7) based on ALKIS function codes
and WSF-derived construction year. Classification logic:
\itemize{
\item Type 7: Bridges (ALKIS 53001_1800)
\item Types 1-3: Residential (ALKIS 31001_1000)
\itemize{
\item Type 1: pre-1986
\item Type 2: 1986-2000
\item Type 3: post-2000
}
\item Types 4-6: Non-residential
\itemize{
\item Type 4: pre-1986
\item Type 5: 1986-2000
\item Type 6: post-2000
}
}
}
\keyword{internal}
