% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_csd_config.R
\name{.find_matching_file}
\alias{.find_matching_file}
\title{Find first file in directory matching pattern}
\usage{
.find_matching_file(pattern, input_dir)
}
\description{
Find first file in directory matching pattern
}
\keyword{internal}
