% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_buildings.R
\name{.force_multipolygon}
\alias{.force_multipolygon}
\title{Normalize geometry to MULTIPOLYGON}
\usage{
.force_multipolygon(geom)
}
\description{
Converts various polygon geometry types (POLYGON, MULTIPOLYGON,
GEOMETRYCOLLECTION) to MULTIPOLYGON format. Returns NULL if
normalization is not possible.
}
\keyword{internal}
