% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download.R
\name{.validate_aoi_projected}
\alias{.validate_aoi_projected}
\title{Validate and project AOI to target EPSG}
\usage{
.validate_aoi_projected(aoi, epsg = 25832)
}
\description{
Validate and project AOI to target EPSG
}
\keyword{internal}
