% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_csd_config.R
\name{.validate_csd_directories}
\alias{.validate_csd_directories}
\title{Validate input and output directories}
\usage{
.validate_csd_directories(output_dir, input_root_path)
}
\description{
Validate input and output directories
}
\keyword{internal}
