% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_csd_config.R
\name{.write_csd_yaml}
\alias{.write_csd_yaml}
\title{Write CSD configuration YAML file}
\usage{
.write_csd_yaml(config_list, output_dir, prefix)
}
\description{
Write CSD configuration YAML file
}
\keyword{internal}
