% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_download.R
\name{download_trees}
\alias{download_trees}
\title{Download tree data for an AOI}
\usage{
download_trees(aoi)
}
\arguments{
\item{aoi}{An sf or sfc polygon object.}
}
\value{
An sf object containing tree points.
}
\description{
Downloads all tree datasets intersecting
an AOI, merges them, and clips the result.
}
