% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsf_download.R
\name{download_wsf_raster}
\alias{download_wsf_raster}
\title{Download, merge, and clip WSF Evolution data to an AOI}
\usage{
download_wsf_raster(aoi)
}
\arguments{
\item{aoi}{An \code{sf} or \code{sfc} object defining the area of interest.}
}
\value{
A \code{terra::SpatRaster} clipped to the AOI.
}
\description{
Downloads all WSF Evolution raster tiles (2×2 degree grid)
intersecting an AOI, mosaics them, and clips the result
to the AOI.
}
\examples{
\dontrun{
library(sf)

aoi <- st_read("inst/extdata/test_aoi.gpkg")

wsf <- download_wsf_data(aoi)

wsf
terra::plot(wsf)
}

}
