% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_raster_stack.R
\name{prepare_raster_stack}
\alias{prepare_raster_stack}
\title{Reproject, clip, and resample rasters to a common grid.}
\usage{
prepare_raster_stack(aoi, target_epsg, resolution, data, out_dir = NULL)
}
\arguments{
\item{aoi}{An \code{sf} or \code{sfc} object defining the area of interest.}

\item{target_epsg}{Integer EPSG code (e.g. 25832).}

\item{resolution}{Target resolution in map units (e.g. meters).}

\item{data}{A named list of \code{terra::SpatRaster} objects.}

\item{out_dir}{Optional output directory. If NULL, nothing is written
to disk and rasters are returned in memory.}
}
\value{
A named list of processed \code{terra::SpatRaster} objects.
}
\description{
The resampling method (bilinear or nearest neighbour) depends on
the name of the \code{terra::SpatRaster} objects in the \emph{named} list.
(LC or WSF --> nearest neighbour, else --> bilinear)
}
