% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings_rasterize.R
\name{rasterize_bridges_palm}
\alias{rasterize_bridges_palm}
\title{Rasterize bridge properties for PALM input}
\usage{
rasterize_bridges_palm(bridges, template)
}
\arguments{
\item{bridges}{An \code{sf} object with bridge polygons and properties.}

\item{template}{A \code{terra::SpatRaster} defining the output grid.}
}
\value{
A named list of two \code{terra::SpatRaster} objects:
\itemize{
\item \code{id}: Bridge ID
\item \code{height}: Bridge height (measuredHeight)
}
}
\description{
Rasterizes bridge ID and height in a single operation.
}
\examples{
\dontrun{
bridges <- res$bridges
rasters <- rasterize_bridges_palm(bridges, dem)

rasters$id
rasters$height
}

}
