% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings_rasterize.R
\name{rasterize_buildings}
\alias{rasterize_buildings}
\title{Rasterize building properties to a regular grid}
\usage{
rasterize_buildings(buildings, template, property, fun = "max")
}
\arguments{
\item{buildings}{An \code{sf} object with building polygons and properties.}

\item{template}{A \code{terra::SpatRaster} defining the output grid.}

\item{property}{The column name to rasterize (e.g., "palm_type", "HoeheGrund").}

\item{fun}{The aggregation function for cells with multiple buildings
(default: "max" - useful for building types and heights).}
}
\value{
A \code{terra::SpatRaster} with building properties rasterized to the grid.
}
\description{
Converts vector building data to raster format, assigning building
properties (type, height, etc.) to grid cells.
}
\examples{
\dontrun{
buildings <- assign_palm_building_type(buildings, wsf)
palm_type_raster <- rasterize_buildings(
  buildings = buildings,
  template = dem,
  property = "palm_type"
)
}

}
