% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_buildings_to_palm.R
\name{rasterize_buildings_to_palm}
\alias{rasterize_buildings_to_palm}
\title{Rasterize all required building properties for PALM input}
\usage{
rasterize_buildings_to_palm(buildings, template)
}
\arguments{
\item{buildings}{An \code{sf} object with building polygons and properties.}

\item{template}{A \code{terra::SpatRaster} defining the output grid.}
}
\value{
A named list of three \code{terra::SpatRaster} objects:
\itemize{
\item \code{type}: Building type classification (palm_type)
\item \code{id}: Building ID
\item \code{height}: Building height (measuredHeight)
}
}
\description{
Rasterizes building type, ID, and height in a single operation.
}
\examples{
\dontrun{
buildings <- assign_palm_building_type(buildings, wsf)
rasters <- rasterize_buildings_palm(buildings, dem)

rasters$type
rasters$id
rasters$height
}

}
