% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_trees_to_palm.R
\name{rasterize_trees_to_palm}
\alias{rasterize_trees_to_palm}
\title{Rasterize tree properties for PALM input}
\usage{
rasterize_trees_to_palm(trees, template, tree_type = 0)
}
\arguments{
\item{trees}{An \code{sf} object containing tree points.}

\item{template}{A \code{terra::SpatRaster} defining the target grid.}

\item{tree_type}{Integer PALM tree type value.}
}
\value{
A named list of rasters.
}
\description{
Rasterizes:
\itemize{
\item tree height
\item tree type
}
}
